Break out ribosome funcs to separate file
